ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.26.3
scikit-learn==1.4.0
detoxify==0.5.2
python-dotenv==1.0.0
httpx==0.26.0
